# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  http.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
...
  http.DestinationImportRow:
    properties:
      id:
        example: 5c657fbd-94f5-4ab8-84c2-dc21b3c21039
        type: string
      job_id:
        example: 35190d3f-1c71-4875-aeaf-d3f8ba0cc905
        type: string
      row_number:
        example: 2
        type: integer
      status:
        enum:
        - pending_review
        - skipped
        - failed
        type: string
      action:
        example: create
        type: string
      change_id:
        example: a468173a-b413-4232-b294-084d47cdd941
        type: string
      error:
        example: destination change validation failed
        type: string
      payload:
        $ref: '#/definitions/http.DestinationChangeFields'
    type: object
  http.DestinationImportJob:
    properties:
      id:
        example: 35190d3f-1c71-4875-aeaf-d3f8ba0cc905
        type: string
      uploaded_by:
        example: 7a13808e-47f9-4af2-b146-8288dd053216
        type: string
      submitted_by:
        example: 7a13808e-47f9-4af2-b146-8288dd053216
        type: string
      submitted_by_full_name:
        type: string
      submitted_by_username:
        type: string
      status:
        enum:
        - queued
        - processing
        - completed
        - failed
        type: string
      dry_run:
        type: boolean
      file_key:
        example: destinations/imports/35190d3f-1c71-4875-aeaf-d3f8ba0cc905/valid_import.csv
        type: string
      error_csv_key:
        type: string
      notes:
        type: string
      total_rows:
        example: 2
        type: integer
      processed_rows:
        example: 2
        type: integer
      rows_failed:
        example: 1
        type: integer
      changes_created:
        example: 1
        type: integer
      pending_change_ids:
        items:
          type: string
        type: array
      submitted_at:
        example: "2025-11-13T17:51:20.243628Z"
        type: string
      completed_at:
        example: "2025-11-13T17:51:20.253166Z"
        type: string
      created_at:
        example: "2025-11-13T17:51:20.243718Z"
        type: string
      updated_at:
        example: "2025-11-13T17:51:20.253259Z"
        type: string
    type: object
  http.DestinationImportResponse:
    properties:
      job:
        $ref: '#/definitions/http.DestinationImportJob'
      rows:
        items:
          $ref: '#/definitions/http.DestinationImportRow'
        type: array
    type: object
...
  http.PageMeta:
    properties:
      count:
        example: 1
        type: integer
      total:
        example: 57
        type: integer
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
    type: object
  http.Destination:
    properties:
      category:
        example: Nature
        type: string
      city:
        example: New York
        type: string
      country:
        example: USA
        type: string
      created_at:
        example: "2024-07-10T12:00:00Z"
        type: string
      deleted_at:
        type: string
      description:
        example: Iconic urban park spanning 843 acres.
        type: string
      contact:
        example: +1-212-310-6600
        type: string
      opening_time:
        example: "08:00"
        type: string
      closing_time:
        example: "22:00"
        type: string
      gallery:
        items:
          $ref: '#/definitions/http.DestinationMedia'
        type: array
      average_rating:
        example: 4.3
        format: double
        type: number
      review_count:
        example: 128
        type: integer
      hero_image_url:
        example: https://cdn.fitcity/destinations/central-park.jpg
        type: string
      id:
        example: 1da5f07d-7dcf-4aea-b948-7c9daa0a5db7
        type: string
      latitude:
        example: 40.785091
        format: double
        type: number
      longitude:
        example: -73.968285
        format: double
        type: number
      name:
        example: Central Park
        type: string
      slug:
        example: central-park
        type: string
      status:
        enum:
        - draft
        - published
        - archived
        type: string
      updated_at:
        example: "2024-07-11T08:00:00Z"
        type: string
      updated_by:
        example: 9fd13fd2-63c5-4f29-a210-4a1a8e285f74
        type: string
      version:
        example: 3
        type: integer
    type: object
  http.DestinationListResponse:
    properties:
      destinations:
        items:
          $ref: '#/definitions/http.Destination'
        type: array
      meta:
        $ref: '#/definitions/http.PageMeta'
    type: object
  http.DestinationResponse:
    properties:
      destination:
        $ref: '#/definitions/http.Destination'
    type: object
  http.DestinationChangeFields:
    description: Editable destination attributes supplied when creating or updating change drafts.
    properties:
      name:
        description: Public display name for the destination.
        example: Central Park
        type: string
      slug:
        description: Optional URL-friendly slug; generated if omitted.
        example: central-park
        type: string
      city:
        description: City where the destination is located.
        example: New York
        type: string
      country:
        description: Country for the destination.
        example: USA
        type: string
      category:
        description: Free-form category label (e.g., Nature, Museum, Food).
        example: Nature
        type: string
      description:
        description: Visitor-facing description or highlights.
        example: Iconic urban park with year-round programming.
        type: string
      latitude:
        description: Decimal latitude for map placement.
        example: 40.785091
        format: double
        type: number
      longitude:
        description: Decimal longitude for map placement.
        example: -73.968285
        format: double
        type: number
      contact:
        description: Optional contact info (phone, email, or booking URL).
        example: "+1 212-310-6600"
        type: string
      opening_time:
        description: Daily opening time in HH:MM (24h) format.
        example: "06:00"
        type: string
      closing_time:
        description: Daily closing time in HH:MM (24h) format.
        example: "22:00"
        type: string
      gallery:
        description: Ordered gallery media to publish with the destination.
        items:
          $ref: '#/definitions/http.DestinationMedia'
        type: array
      hero_image_upload_id:
        description: Temporary upload identifier returned from the hero upload endpoint.
        example: 9c2d4ee5-32be-4f08-b4c0-f59196c6f36f
        type: string
      hero_image_url:
        description: Fully-qualified URL to an existing hero image when reusing published media.
        example: https://cdn.fitcity/destinations/central-park.jpg
        type: string
      published_hero_image:
        description: (Read-only) currently published hero image URL; useful for draft diffs.
        example: https://cdn.fitcity/destinations/central-park.jpg
        type: string
      hard_delete:
        description: When true and action is delete, removes the destination permanently.
        type: boolean
      status:
        description: Desired publication status when approving the draft.
        enum:
        - draft
        - published
        - archived
        type: string
    type: object
  http.DestinationMedia:
    properties:
      url:
        description: Publicly accessible image URL or CDN path.
        example: https://cdn.fitcity/destinations/central-park/gallery-1.jpg
        type: string
      caption:
        description: Short description rendered under the image.
        example: Summer blooms near the lake
        type: string
      ordering:
        description: Zero-based ordering for display.
        example: 1
        type: integer
    required:
    - url
    type: object
...
info:
  contact: {}
  description: Authentication and profile services for the Fit City platform.
  title: Fit City API
  version: "1.0"
paths:
...
  /destinations:
    get:
      description: List published destinations visible to end users.
      parameters:
      - description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      - description: Number of records to skip
        in: query
        name: offset
        type: integer
      - description: Full text search across name, city, country, category, or description
        in: query
        name: query
        type: string
      - description: Comma-separated list of categories to include
        in: query
        name: categories
        type: string
      - description: Minimum average rating (0-5)
        format: double
        in: query
        name: min_rating
        type: number
      - description: Maximum average rating (0-5)
        format: double
        in: query
        name: max_rating
        type: number
      - description: Country filter
        in: query
        name: country
        type: string
      - description: City filter
        in: query
        name: city
        type: string
      - description: Latitude for nearby search
        in: query
        name: lat
        type: number
        format: double
      - description: Longitude for nearby search
        in: query
        name: lng
        type: number
        format: double
      - description: Max distance in kilometers for nearby search
        in: query
        name: max_distance_km
        type: number
        format: double
      - description: Additional Sort (similarity, distance)
        in: query
        name: sort
        type: string
      - description: Sort order (rating_desc, rating_asc, alpha_asc, alpha_desc, updated_at_desc)
        in: query
        name: sort
        type: strings
      - description: Sort order (rating_desc, rating_asc, alpha_asc, alpha_desc, updated_at_desc)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DestinationListResponse'
      summary: List published destinations
      tags:
      - Destinations
...
  /destinations/{id}:
    get:
      description: Retrieve a published destination by UUID or slug.
      parameters:
      - description: Destination UUID or slug
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DestinationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get published destination
      tags:
      - Destinations
...
  /admin/destination-imports/template:
    get:
      description: Download a CSV template with headers and a sample row for destination imports.
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get destination import template
      tags:
      - Admin Destinations
  /admin/destination-imports:
    post:
      consumes:
      - multipart/form-data
      description: Upload a CSV file to create destination drafts in bulk. Use the dry_run flag to validate without persisting change requests.
      parameters:
      - description: When true, validates the CSV without creating change requests.
        in: query
        name: dry_run
        type: boolean
      - description: CSV file containing destination rows.
        in: formData
        name: file
        required: true
        type: file
      - description: Optional operator notes stored with the job.
        in: formData
        name: notes
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/http.DestinationImportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload destination import CSV
      tags:
      - Admin Destinations
  /admin/destination-imports/{id}:
    get:
      description: Fetch a destination import job and its row-level details.
      parameters:
      - description: Import job UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.DestinationImportResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get destination import job
      tags:
      - Admin Destinations
  /admin/destination-imports/{id}/errors:
    get:
      description: Download a CSV describing all failed rows for a given import job.
      parameters:
      - description: Import job UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/csv
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download destination import errors
      tags:
      - Admin Destinations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"